requirements.txt
fhir.resources~=7.0.2
psycopg2-binary~=2.9.7
python-dateutil~=2.8.2
requests~=2.31.0
